+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Show selected issue details immediately while loading comments (`vsbeads-jw3`)
+
 ## [0.14.1] - 2026-07-30
 
 ### Fixed